Eliminate flicker/slowdown on Home2 style.
When one or more items in the Home2 wraplist are hidden, the left-most item
in the list would flicker when scrolling right.  One user on the forums also
reported massive framerate drop in this condition.

The (temporary) fix appears to be to set preloaditems to 0 (zero).  This
fixes the flickering that I've personally seen.  Although I don't know if it
fixes the slowdown as well since I can't reproduce that here, I suspect that
it probably will.

For reference, see:

http://trac.xbmc.org/ticket/12145
https://github.com/stoli/Metropolis/issues/21