Remove mismatched/unnecessary icons.
Many icons in the Files views were mismatched.  Instead of trying to add
custom icons for everything in sight, it's easier to just remove the few
custom icons there are to make everything match.

Would have liked to remove DefaultActor.png and DefaultArtist.png as well
but the fallbacks are not working in XBMC core.  Confluence does this the
same way, so I suspect this is a known issue.